Encapsulate the game state management into a shared class
