+++ games/AirHockey/AirHockey.Server/Program.cs
@@ -1,11 +1,14 @@
 using AirHockey.Server;
 using AirHockey.Server.Hubs;
+using AirHockey.Shared;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddHostedService<GameWorker>();
 
 builder.Services.AddSingleton<GameManager>();
+builder.Services.AddTransient<GameStateManager>();
+builder.Services.AddTransient<IGameLifeCycleHandler, SignalRGameLifeCycleHandler>();
 
 builder.Services.AddSignalR();
 
@@ -13,29 +16,6 @@ var app = builder.Build();
 
 app.UseHttpsRedirection();
 
-//app.MapGet("/", () => "Hello World!");
-
 app.MapHub<GameHub>("Game");
 
 app.Run();
-
-
-// using DrawingGame.Server.Hubs;
-
-// var builder = WebApplication.CreateBuilder(args);
-
-// // Add services to the container.
-// builder.Services.AddSignalR();
-
-// var app = builder.Build();
-
-// // Configure the HTTP request pipeline.
-// if (app.Environment.IsDevelopment())
-// {
-// }
-
-// app.UseHttpsRedirection();
-
-// app.MapHub<GameHub>("Game");
-
-// app.Run();
